Agents/research-assistant/requirements.txt
numpy
requests
python-dotenv
feedparser
PyMuPDF
tiktoken
sendgrid
pinecone-client
pytest
sentence-transformers
ollama
